Refactor import paths for RECURRENCE to use centralized constants location
